Refactor server tenant registry module
